Move vendor and risk markdown fun to markdown folder
