+++ Servers/services/reportService.ts
@@ -1,9 +1,7 @@
-import { getProjectRisksReportQuery, getMembersByProjectIdQuery } from "../utils/reporting.utils";
+import { getMembersByProjectIdQuery } from "../utils/reporting.utils";
 import { DefaultReportName, ReportType } from "../models/reporting.model";
-import { getVendorRisksByProjectIdQuery } from "../utils/vendorRisk.util";
-import { getVendorByProjectIdQuery } from "../utils/vendor.utils";
-import { getUserByIdQuery } from "../utils/user.utils";
 import { getProjectRiskMarkdown } from "./markdowns/projectRiskMarkdown";
+import { getVendorReportMarkdown } from "./markdowns/vendorAndRisksMarkdown";
 import { getAssessmentTrackerMarkdown } from './markdowns/assessmentTrackerMarkdown';
 
 export interface reportBodyData {
@@ -104,62 +102,3 @@ export async function getReportData(
   
 
 
-/**
- * Generates a markdown report for vendor risks
- * @param projectId - The ID of the project
- * @param data - Project metadata including title and owner
- * @returns Promise<string> - Markdown formatted vendor risk report
- */
-export async function getVendorReportMarkdown(
-  projectId: number,
-  data: reportBodyData
-): Promise<any> {
-  const vendorData = await getVendorByProjectIdQuery(projectId);
-  const riskData = await getVendorRisksByProjectIdQuery(projectId);
-
-  const vendorRows: string =
-    vendorData && vendorData.length > 0
-      ? await Promise.all(
-          vendorData.map(async (vendor) => {
-            const vendorAssignee = await getUserByIdQuery(vendor.assignee);
-            return `| ${vendor.vendor_name} | ${vendorAssignee.name} ${
-              vendorAssignee.surname
-            } | ${vendor.review_status} | ${
-              vendor.risk_status
-            } | ${vendor.review_date.toLocaleDateString()} |`;
-          })
-        ).then((rows) => rows.join("\n"))
-      : "| - | - | - | - | - |";
-
-  const riskRows: string =
-    riskData && riskData.length > 0
-      ? await Promise.all(
-          riskData.map(async (risk) => {
-            const riskActionOwner = await getUserByIdQuery(risk.action_owner);
-            return `| ${risk.impact} | ${risk.likelihood} | ${risk.risk_severity} | ${riskActionOwner.name} ${riskActionOwner.surname} | ${risk.risk_level} | ${risk.risk_description} |`;
-          })
-        ).then((rows) => rows.join("\n"))
-      : "| - | - | - | - | - | - |";
-
-  return `
-VerifyWise vendors and vendor risks report
--------------
-This report is generated by VerifyWise. It aims to give information about the company's vendors and corresponding vendor risks.
-  
-- **Date**: ${new Date().toLocaleDateString()}
-- **Project**: ${data.projectTitle}
-- **Owner**: ${data.projectOwner}
-
-Vendors table
--------------
-| Vendor Name | Assignee | Review Status | Risk Status | Review Date |
-|----|----|----|----|----|
-${vendorRows}
-
-Vendor risk table
--------------
-| Impact | Likelihood | Risk Severity | Action Owner | Risk Level | Risk Description |
-|----|----|----|----|----|----|
-${riskRows}
-`;
-}
